Update filter and its interface
